requirements_dev.txt
flake8==4.0.1
mypy==0.961
types-requests==2.27.20
pytest==7.1.2
pytest-cov==3.0.0
vcrpy==4.1.1
python-dotenv==0.20.0
pre-commit==2.19.0
